+++ imessage-database/src/tables/messages/message.rs
@@ -622,6 +622,41 @@ impl Message {
         0
     }
 
+    /// Generate the SQL `WHERE` clause described by a [`QueryContext`]
+    pub(crate) fn generate_filter_statement(context: &QueryContext) -> String {
+        let mut filters = String::new();
+        if let Some(start) = context.start {
+            filters.push_str(&format!("    m.date >= {start}"));
+        }
+        if let Some(end) = context.end {
+            if !filters.is_empty() {
+                filters.push_str(" AND ");
+            }
+            filters.push_str(&format!("    m.date <= {end}"));
+        }
+        if let Some(chat_ids) = &context.selected_handle_ids {
+            if !filters.is_empty() {
+                filters.push_str(" AND ");
+            }
+            filters.push_str(&format!(
+                "    m.chat_id IN ({})",
+                chat_ids
+                    .iter()
+                    .map(|x| x.to_string())
+                    .collect::<Vec<String>>()
+                    .join(", "),
+            ));
+        }
+
+        if !filters.is_empty() {
+            return format!(
+                " WHERE
+                 {filters}"
+            );
+        }
+        filters
+    }
+
     /// Get the number of messages in the database
     ///
     /// # Example:
@@ -641,7 +676,7 @@ impl Message {
         let mut statement = if context.has_filters() {
             db.prepare(&format!(
                 "SELECT COUNT(*) FROM {MESSAGE} as m {}",
-                context.generate_filter_statement("m.date")
+                Self::generate_filter_statement(context)
             ))
             .map_err(TableError::Messages)?
         } else {
@@ -676,7 +711,7 @@ impl Message {
             return Self::get(db);
         }
 
-        let filters = context.generate_filter_statement("m.date");
+        let filters = Self::generate_filter_statement(context);
 
         // If database has `thread_originator_guid`, we can parse replies, otherwise default to 0
         Ok(db.prepare(&format!(
@@ -712,7 +747,7 @@ impl Message {
     }
 
     /// See [`Tapback`] for details on this data.
-    fn clean_associated_guid(&self) -> Option<(usize, &str)> {
+    pub(crate) fn clean_associated_guid(&self) -> Option<(usize, &str)> {
         if let Some(guid) = &self.associated_message_guid {
             if guid.starts_with("p:") {
                 let mut split = guid.split('/');
@@ -828,7 +863,7 @@ impl Message {
     ///
     /// For example, a Bundle ID like `com.apple.messages.MSMessageExtensionBalloonPlugin:0000000000:com.apple.SafetyMonitorApp.SafetyMonitorMessages`
     /// should get parsed into `com.apple.SafetyMonitorApp.SafetyMonitorMessages`.
-    fn parse_balloon_bundle_id(&self) -> Option<&str> {
+    pub(crate) fn parse_balloon_bundle_id(&self) -> Option<&str> {
         if let Some(bundle_id) = &self.balloon_bundle_id {
             let mut parts = bundle_id.split(':');
             let bundle_id = parts.next();
@@ -1064,18 +1099,8 @@ impl Message {
 }
 
 #[cfg(test)]
-mod tests {
-    use crate::{
-        message_types::{
-            edited::{EditStatus, EditedMessage, EditedMessagePart},
-            expressives,
-            variants::{CustomBalloon, Variant},
-        },
-        tables::messages::Message,
-        util::dates::get_offset,
-    };
-
-    fn blank() -> Message {
+impl Message {
+    pub fn blank() -> Message {
         Message {
             rowid: i32::default(),
             guid: String::default(),
@@ -1110,299 +1135,5 @@ mod tests {
             components: None,
             edited_parts: None,
         }
-    }
-
-    #[test]
-    fn can_gen_message() {
-        blank();
-    }
-
-    #[test]
-    fn can_get_time_date_read_after_date() {
-        // Get offset
-        let offset = get_offset();
-
-        // Create message
-        let mut message = blank();
-        // May 17, 2022  8:29:42 PM
-        message.date = 674526582885055488;
-        // May 17, 2022  8:29:42 PM
-        message.date_delivered = 674526582885055488;
-        // May 17, 2022  9:30:31 PM
-        message.date_read = 674530231992568192;
-
-        assert_eq!(
-            message.time_until_read(&offset),
-            Some("1 hour, 49 seconds".to_string())
-        );
-    }
-
-    #[test]
-    fn can_get_time_date_read_before_date() {
-        // Get offset
-        let offset = get_offset();
-
-        // Create message
-        let mut message = blank();
-        // May 17, 2022  9:30:31 PM
-        message.date = 674530231992568192;
-        // May 17, 2022  9:30:31 PM
-        message.date_delivered = 674530231992568192;
-        // May 17, 2022  8:29:42 PM
-        message.date_read = 674526582885055488;
-
-        assert_eq!(message.time_until_read(&offset), None);
-    }
-
-    #[test]
-    fn can_get_message_expression_none() {
-        let m = blank();
-        assert_eq!(m.get_expressive(), expressives::Expressive::None);
-    }
-
-    #[test]
-    fn can_get_message_expression_bubble() {
-        let mut m = blank();
-        m.expressive_send_style_id = Some("com.apple.MobileSMS.expressivesend.gentle".to_string());
-        assert_eq!(
-            m.get_expressive(),
-            expressives::Expressive::Bubble(expressives::BubbleEffect::Gentle)
-        );
-    }
-
-    #[test]
-    fn can_get_message_expression_screen() {
-        let mut m = blank();
-        m.expressive_send_style_id =
-            Some("com.apple.messages.effect.CKHappyBirthdayEffect".to_string());
-        assert_eq!(
-            m.get_expressive(),
-            expressives::Expressive::Screen(expressives::ScreenEffect::Balloons)
-        );
-    }
-
-    #[test]
-    fn can_get_no_balloon_bundle_id() {
-        let m = blank();
-        assert_eq!(m.parse_balloon_bundle_id(), None);
-    }
-
-    #[test]
-    fn can_get_balloon_bundle_id_os() {
-        let mut m = blank();
-        m.balloon_bundle_id = Some("com.apple.Handwriting.HandwritingProvider".to_owned());
-        assert_eq!(
-            m.parse_balloon_bundle_id(),
-            Some("com.apple.Handwriting.HandwritingProvider")
-        );
-    }
-
-    #[test]
-    fn can_get_balloon_bundle_id_url() {
-        let mut m = blank();
-        m.balloon_bundle_id = Some("com.apple.messages.URLBalloonProvider".to_owned());
-        assert_eq!(
-            m.parse_balloon_bundle_id(),
-            Some("com.apple.messages.URLBalloonProvider")
-        );
-    }
-
-    #[test]
-    fn can_get_balloon_bundle_id_apple() {
-        let mut m = blank();
-        m.balloon_bundle_id = Some("com.apple.messages.MSMessageExtensionBalloonPlugin:0000000000:com.apple.PassbookUIService.PeerPaymentMessagesExtension".to_owned());
-        assert_eq!(
-            m.parse_balloon_bundle_id(),
-            Some("com.apple.PassbookUIService.PeerPaymentMessagesExtension")
-        );
-    }
-
-    #[test]
-    fn can_get_balloon_bundle_id_third_party() {
-        let mut m = blank();
-        m.balloon_bundle_id = Some("com.apple.messages.MSMessageExtensionBalloonPlugin:QPU8QS3E62:com.contextoptional.OpenTable.Messages".to_owned());
-        assert_eq!(
-            m.parse_balloon_bundle_id(),
-            Some("com.contextoptional.OpenTable.Messages")
-        );
-        assert!(matches!(
-            m.variant(),
-            Variant::App(CustomBalloon::Application(
-                "com.contextoptional.OpenTable.Messages"
-            ))
-        ));
-    }
-
-    #[test]
-    fn can_get_valid_guid() {
-        let mut m = blank();
-        m.associated_message_guid = Some("A44CE9D7-AAAA-BBBB-CCCC-23C54E1A9B6A".to_string());
-
-        assert_eq!(
-            Some((0usize, "A44CE9D7-AAAA-BBBB-CCCC-23C54E1A9B6A")),
-            m.clean_associated_guid()
-        );
-    }
-
-    #[test]
-    fn cant_get_invalid_guid() {
-        let mut m = blank();
-        m.associated_message_guid = Some("FAKE_GUID".to_string());
-
-        assert_eq!(None, m.clean_associated_guid());
-    }
-
-    #[test]
-    fn can_get_valid_guid_p() {
-        let mut m = blank();
-        m.associated_message_guid = Some("p:1/A44CE9D7-AAAA-BBBB-CCCC-23C54E1A9B6A".to_string());
-
-        assert_eq!(
-            Some((1usize, "A44CE9D7-AAAA-BBBB-CCCC-23C54E1A9B6A")),
-            m.clean_associated_guid()
-        );
-    }
-
-    #[test]
-    fn cant_get_invalid_guid_p() {
-        let mut m = blank();
-        m.associated_message_guid = Some("p:1/FAKE_GUID".to_string());
-
-        assert_eq!(None, m.clean_associated_guid());
-    }
-
-    #[test]
-    fn can_get_valid_guid_bp() {
-        let mut m = blank();
-        m.associated_message_guid = Some("bp:A44CE9D7-AAAA-BBBB-CCCC-23C54E1A9B6A".to_string());
-
-        assert_eq!(
-            Some((0usize, "A44CE9D7-AAAA-BBBB-CCCC-23C54E1A9B6A")),
-            m.clean_associated_guid()
-        );
-    }
-
-    #[test]
-    fn cant_get_invalid_guid_bp() {
-        let mut m = blank();
-        m.associated_message_guid = Some("bp:FAKE_GUID".to_string());
-
-        assert_eq!(None, m.clean_associated_guid());
-    }
-
-    #[test]
-    fn can_get_fully_unsent_true_single() {
-        let mut m = blank();
-        m.edited_parts = Some(EditedMessage {
-            parts: vec![EditedMessagePart {
-                status: EditStatus::Unsent,
-                edit_history: vec![],
-            }],
-        });
-
-        assert!(m.is_fully_unsent());
-    }
-
-    #[test]
-    fn can_get_fully_unsent_true_multiple() {
-        let mut m = blank();
-        m.edited_parts = Some(EditedMessage {
-            parts: vec![
-                EditedMessagePart {
-                    status: EditStatus::Unsent,
-                    edit_history: vec![],
-                },
-                EditedMessagePart {
-                    status: EditStatus::Unsent,
-                    edit_history: vec![],
-                },
-            ],
-        });
-
-        assert!(m.is_fully_unsent());
-    }
-
-    #[test]
-    fn can_get_fully_unsent_false() {
-        let mut m = blank();
-        m.edited_parts = Some(EditedMessage {
-            parts: vec![EditedMessagePart {
-                status: EditStatus::Original,
-                edit_history: vec![],
-            }],
-        });
-
-        assert!(!m.is_fully_unsent());
-    }
-
-    #[test]
-    fn can_get_fully_unsent_false_multiple() {
-        let mut m = blank();
-        m.edited_parts = Some(EditedMessage {
-            parts: vec![
-                EditedMessagePart {
-                    status: EditStatus::Unsent,
-                    edit_history: vec![],
-                },
-                EditedMessagePart {
-                    status: EditStatus::Original,
-                    edit_history: vec![],
-                },
-            ],
-        });
-
-        assert!(!m.is_fully_unsent());
-    }
-
-    #[test]
-    fn can_get_part_edited_true() {
-        let mut m = blank();
-        m.edited_parts = Some(EditedMessage {
-            parts: vec![
-                EditedMessagePart {
-                    status: EditStatus::Edited,
-                    edit_history: vec![],
-                },
-                EditedMessagePart {
-                    status: EditStatus::Original,
-                    edit_history: vec![],
-                },
-            ],
-        });
-
-        assert!(m.is_part_edited(0));
-    }
-
-    #[test]
-    fn can_get_part_edited_false() {
-        let mut m = blank();
-        m.edited_parts = Some(EditedMessage {
-            parts: vec![
-                EditedMessagePart {
-                    status: EditStatus::Edited,
-                    edit_history: vec![],
-                },
-                EditedMessagePart {
-                    status: EditStatus::Original,
-                    edit_history: vec![],
-                },
-            ],
-        });
-
-        assert!(!m.is_part_edited(1));
-    }
-
-    #[test]
-    fn can_get_part_edited_blank() {
-        let m = blank();
-
-        assert!(!m.is_part_edited(0));
-    }
-
-    #[test]
-    fn can_get_fully_unsent_none() {
-        let m = blank();
-
-        assert!(!m.is_fully_unsent());
     }
 }
