+++ imessage-database/src/tables/messages/tests/mod.rs
@@ -1,0 +1,6 @@
+mod bundle_id_tests;
+mod date_tests;
+mod edited_tests;
+mod expressive_tests;
+mod guid_tests;
+mod query_tests;
